Move writer config to private module
Now that the class constructor is private, the config object passed to
the constructor can also be private. Unlike the class, however, this
doesn't need any special usage -- just an ordinary private name.

In order to keep the project layout tidy, this is grouped with the writer
itself in a subpackage.
